selenium-python-framework-structure/requirements.txt
webdriver_manager
selenium~=4.1.3
pytest~=6.0.1
pytest-bdd~=3.4.0
pytest-html
python-dateutil~=2.8.1
configparser~=5.0.2

sphinx
sphinx_rtd_theme
sphinx-autopackagesummary

allure-pytest==2.8.13
assertpy
PyPDF4
Appium-Python-Client~=2.2.0